refactored TestUtil and integration tests
