+++ vsphere/src/test/java/com/elastisys/scale/cloudpool/vsphere/IntegrationTestTagger.java
@@ -1,8 +1,6 @@
 package com.elastisys.scale.cloudpool.vsphere;
 
 import com.elastisys.scale.cloudpool.commons.basepool.driver.DriverConfig;
-import com.elastisys.scale.cloudpool.vsphere.client.VsphereClient;
-import com.elastisys.scale.cloudpool.vsphere.client.impl.StandardVsphereClient;
 import com.elastisys.scale.cloudpool.vsphere.tag.Tag;
 import com.elastisys.scale.cloudpool.vsphere.tag.impl.VsphereTag;
 import com.elastisys.scale.cloudpool.vsphere.client.tagger.Tagger;
@@ -10,7 +8,6 @@ import com.elastisys.scale.cloudpool.vsphere.client.tagger.impl.CustomAttributeT
 import com.elastisys.scale.cloudpool.vsphere.driver.config.VsphereApiSettings;
 import com.elastisys.scale.cloudpool.vsphere.driver.config.VsphereProvisioningTemplate;
 import com.elastisys.scale.commons.json.JsonUtils;
-import com.google.common.collect.Lists;
 import com.vmware.vim25.VirtualMachineCloneSpec;
 import com.vmware.vim25.VirtualMachineRelocateSpec;
 import com.vmware.vim25.mo.*;
@@ -19,13 +16,11 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.rmi.RemoteException;
-import java.util.List;
 import java.util.UUID;
 
-import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.*;
 
-public class ITVsphere {
+public class IntegrationTestTagger {
 
     private static DriverConfig driverConfig;
     private static VsphereApiSettings vsphereApiSettings;
@@ -81,32 +76,6 @@ public class ITVsphere {
         assertTrue(tagger.isTagged(minimalVm, tag));
         tagger.untag(minimalVm, tag);
         assertFalse(tagger.isTagged(minimalVm, tag));
-    }
-
-    @Test
-    public void clientConfigurationShouldSucceed() throws RemoteException {
-        VsphereClient vsphereClient = new StandardVsphereClient();
-        vsphereClient.configure(vsphereApiSettings, vsphereProvisioningTemplate);
-    }
-
-    @Test
-    public void shouldNotGetVirtualMachinesForTagThatDoesNotExist() throws RemoteException {
-        VsphereClient vsphereClient = new StandardVsphereClient();
-        vsphereClient.configure(vsphereApiSettings, vsphereProvisioningTemplate);
-        List<Tag> tags = Lists.newArrayList();
-        tags.add(new VsphereTag(VsphereTag.ScalingTag.CLOUD_POOL, "NoSuchPool"));
-        List<VirtualMachine> virtualMachines = vsphereClient.getVirtualMachines(tags);
-        assertTrue(virtualMachines.isEmpty());
-    }
-
-    @Test
-    public void shouldGetVirtualMachineWithoutTagRequirements() throws RemoteException {
-        // this test assumes that there is at least one virtual machine or template ManagedEntity on the server
-        VsphereClient vsphereClient = new StandardVsphereClient();
-        vsphereClient.configure(vsphereApiSettings, vsphereProvisioningTemplate);
-        List<Tag> tags = Lists.newArrayList();
-        List<VirtualMachine> virtualMachines = vsphereClient.getVirtualMachines(tags);
-        assertFalse(virtualMachines.isEmpty());
     }
 
 }
