+++ vsphere/src/test/java/com/elastisys/scale/cloudpool/vsphere/client/impl/TestStandardVsphereClient.java
@@ -1,12 +1,10 @@
 package com.elastisys.scale.cloudpool.vsphere.client.impl;
 
 import com.elastisys.scale.cloudpool.commons.basepool.driver.DriverConfig;
-import com.elastisys.scale.cloudpool.vsphere.TestUtils;
 import com.elastisys.scale.cloudpool.vsphere.driver.config.VsphereApiSettings;
 import com.elastisys.scale.cloudpool.vsphere.driver.config.VsphereProvisioningTemplate;
 
-import com.elastisys.scale.cloudpool.vsphere.tag.Tag;
-import com.elastisys.scale.cloudpool.vsphere.tag.impl.VsphereTag;
+import com.elastisys.scale.commons.json.JsonUtils;
 import com.google.common.collect.Lists;
 import com.vmware.vim25.mo.*;
 import org.junit.Before;
@@ -17,7 +15,6 @@ import java.net.URL;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
 import static org.mockito.Mockito.*;
 
 public class TestStandardVsphereClient {
@@ -31,7 +28,7 @@ public class TestStandardVsphereClient {
 
     @BeforeClass
     public static void setUpBeforeClass() throws Exception {
-        DriverConfig driverConfig = TestUtils.loadDriverConfig("config/valid-vsphere-config.json");
+        DriverConfig driverConfig = JsonUtils.toObject(JsonUtils.parseJsonResource("config/valid-vsphere-config.json"), DriverConfig.class);
         vsphereApiSettings = driverConfig.parseCloudApiSettings(VsphereApiSettings.class);
         vsphereProvisioningTemplate = driverConfig.parseProvisioningTemplate(VsphereProvisioningTemplate.class);
         vsphereClient = spy(new StandardVsphereClient());
